+++ checkpoint/orbax/checkpoint/pytree_checkpoint_handler.py
@@ -610,7 +610,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
       self,
       aggregate_filename: Optional[str] = None,
       concurrent_gb: int = 96,
-      use_ocdbt: bool = False,
+      use_ocdbt: bool = True,
       restore_with_serialized_types: bool = True,
       write_tree_metadata: bool = False,
   ):