+++ checkpoint/CHANGELOG.md
@@ -22,6 +22,7 @@ within `async` functions.
 - Modified `StringHandler` serialization and deserialization to use Tensorstore
 Json driver for async file reads and writes.
 - Marked `transform_utils` as deprecated.
+- Change `PytreeCheckpointHandler`'s parameter `use_ocdbt` default to `True`
 
 ## [0.3.5] - 2022-08-17